Moved custom element registration into a decorator. Rolled target back from ESNext to ES2024 for browser compatability with decorators. Implemented a typed createElement method for custom elements. Added doc comments
