+++ .prettierrc
@@ -1,5 +1,5 @@
 {
-	"plugins": ["@trivago/prettier-plugin-sort-imports"],
+	"plugins": ["decorators", "@trivago/prettier-plugin-sort-imports"],
 	"arrowParens": "avoid",
 	"bracketSameLine": false,
 	"bracketSpacing": true,
