+++ src/scripts/supplement/elements.ts
@@ -1,5 +1,3 @@
-import { registerCustomElements } from "../components/index.js";
-
 /**
  * Constructor signature of a {@link HTMLElement} or any children.
  *
@@ -7,6 +5,25 @@ import { registerCustomElements } from "../components/index.js";
  * WebComponents.
  */
 export type ElementConstructor<E extends HTMLElement> = new () => E;
+
+/**
+ * Create an instance of a custom element of a specified class.
+ *
+ * @param constructor The constructor for the class of the element to create.
+ * @returns The created element.
+ */
+export function createCustomElement<const CustomElement extends typeof HTMLElement>(
+	constructor: CustomElement
+): InstanceType<CustomElement> {
+	// ensure tag for custom element is registered
+	const tagName = customElements.getName(constructor);
+	if (tagName === null) throw ElementError.nonExistantCustomElement(constructor);
+
+	// create custom element and ensure class is correct
+	const element = document.createElement(tagName);
+	if (element instanceof constructor) return element as InstanceType<typeof constructor>;
+	else throw ElementError.incorrectCreatedInstanceType(tagName, constructor, element);
+}
 
 /**
  * Retrieve an element from the DOM, ensuring it is of an expected class.
@@ -22,9 +39,6 @@ export type ElementConstructor<E extends HTMLElement> = new () => E;
  * @returns The element if successfully found.
  */
 export function getElement<E extends HTMLElement>(id: string, constructor: ElementConstructor<E>) {
-	// ensure WebComponents have been registered
-	registerCustomElements();
-
 	// find element
 	const element = document.getElementById(id);
 	if (element === null) throw ElementError.notFound(id);
@@ -63,6 +77,41 @@ class ElementError extends Error {
 	) {
 		return new ElementError(
 			`Element with ID '${id}' has tag '${element.tagName.toLowerCase()}' and is not an instance of '${constructor.name}'.`
+		);
+	}
+
+	/**
+	 * Construct an error for when a custom element is created without being registered.
+	 * 
+	 * @param constructor The element that was requested to be created.
+	 * @returns The constructed {@link ElementError}.
+	 */
+	static nonExistantCustomElement(constructor: typeof HTMLElement) {
+		return new ElementError(
+			`'${constructor.name}' as not been registered as a custom element and cannot be initialised.`
+		);
+	}
+
+	/**
+	 * Construct an error for when a WebComponent was created with an incorrect class.
+	 *
+	 * @param tagName The tag used to create the element.
+	 * @param constructor The constructor which should have been used to create the WebComponent.
+	 * @param element The element that was created.
+	 * @returns The constructed {@link ElementError}.
+	 */
+	static incorrectCreatedInstanceType(
+		tagName: string,
+		constructor: ElementConstructor<HTMLElement>,
+		element: HTMLElement
+	) {
+		// attempt to extract class name from element
+		let className = element.toString();
+		const matches = /\[object (.+?)\]/.exec(className);
+		if (matches !== null) className = matches[1];
+
+		return new ElementError(
+			`WebComponent '${tagName.toLowerCase()}' did not create element with prototype of '${constructor.name}', instead created '${className}'.`
 		);
 	}
 }
